chore: move "Automatable" from parameters.ts to graph.ts
